Move sensor data back into the platform-file
